requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
torch==2.1.2
torchvision==0.16.2
transformers==4.36.2
pillow==10.1.0
pydantic==2.5.2
pydantic-settings==2.1.0
numpy==1.24.3
